Split the agent runtime by what it decides
The file had reached 2154 lines holding six concerns, past the point its
own module doc described: that doc said the bulky separable parts lived
in submodules, which had stopped being true.

The seams are what gets read on its own. Report owns the schema Gemini
is asked for and the validation deciding whether a response is one,
which belong together because a schema nobody validates against is a
suggestion. Events owns the four topic appliers and their dispatcher,
which share the runtime state and the rule that a browser payload is a
claim rather than a fact. Value is the JSON to Python coercion, which is
not about interviews at all and was sitting at the bottom of the file.

Everything is re-exported, so no caller changed and all 69 items are the
same code in a different file. Three functions lost visibility that
nothing outside their own module was using.
