+++ src/agent.rs
@@ -1,14 +1,25 @@
 //! Interview runtime: session state, timing decisions, and report shaping.
 //!
-//! The bulky, separable concerns live in submodules and are re-exported here,
-//! so callers keep importing from `crate::agent` and this file stays about
-//! behavior rather than data.
+//! What is left here is the state an interview carries and the decisions taken
+//! against it: `RuntimeState`, the framework evidence recorded into it, the
+//! timing rules, and the interview configuration a room is started with.
+//!
+//! The separable concerns are submodules, re-exported here so callers keep
+//! importing from `crate::agent`. They are separate because each is read on its
+//! own: `events` decides what to believe from the candidate's browser, `report`
+//! owns the schema Gemini is asked for and whether a response is one, `value`
+//! coerces JSON the way the browser and Python actually behave, and `problems`
+//! and `prompts` are the data this file used to hold.
 
+mod events;
 mod integrity;
 mod problem_topics;
 mod problems;
 mod prompts;
+mod report;
+mod value;
 
+pub use events::apply_data_event;
 use integrity::integrity_hash;
 pub use integrity::{sanitize_integrity_event, sanitize_test_run};
 pub use problems::{DEFAULT_PROBLEM_ID, PROBLEMS, get_problem, topics_for};
@@ -18,9 +29,14 @@ pub use prompts::{
     report_prompt, significant_change, silence_nudge, spoken_language, test_results_reaction,
     time_warning, wrap_up,
 };
+pub use report::{
+    fallback_report, final_report, report_response_schema, validate_report,
+    validate_report_candidate,
+};
+pub use value::json_number;
+pub(crate) use value::{json_int, python_truthy, truthy_string, value_string};
 
 use crate::config::{DEFAULT_DURATION_MIN, MAX_DURATION_MIN, MIN_DURATION_MIN};
-use crate::runtime::{TOPIC_CODE_UPDATE, TOPIC_CONTROL, TOPIC_INTEGRITY, TOPIC_TEST_RESULTS};
 
 pub const MAX_INTEGRITY_EVENTS: usize = 25;
 /// The `detail` bound, and half of a wire contract: `INTEGRITY_DETAIL_MAX` in
@@ -910,626 +926,6 @@ pub fn format_transcript(items: &[TranscriptItem<'_>]) -> String {
         .join("\n")
 }
 
-/// An evaluation that could not be produced is not an evaluation of zero.
-///
-/// This used to emit `codingScore: 0`, `communicationScore: 0` and
-/// `decision: "NO_HIRE"`, so a Gemini outage reached the candidate as a
-/// rejection with 0/100 on both axes, and was then saved to their history and
-/// synced to `/api/reports`. The `error: true` flag it set alongside was
-/// dropped by the browser's own sanitizer, so nothing downstream could tell the
-/// difference between a failed provider and a failed candidate. `incomplete`
-/// is the shape the browser already renders as "No evaluation", with no
-/// verdict, no scores and no badge, and it is the only flag: an `error: true`
-/// that travelled beside it reached no consumer and was a second name for the
-/// same fact.
-pub fn fallback_report(hints_used: u32, note: &str) -> serde_json::Value {
-    let note = note.chars().take(240).collect::<String>();
-    serde_json::json!({
-        "incomplete": true,
-        "summary": format!(
-            "The automatic evaluation could not be completed: {note}. Your session ran end-to-end, but no scores were produced, so nothing here is an assessment of your work. Check the agent logs and GOOGLE_API_KEY, then try again."
-        ),
-        "improvementPlan": [],
-        "frameworkAssessment": null,
-        "hintsUsed": hints_used,
-    })
-}
-
-pub fn report_response_schema() -> serde_json::Value {
-    // `responseSchema` accepts only Gemini's OpenAPI subset, which has no
-    // `additionalProperties`: sending it fails the whole call with a 400 naming
-    // an unknown field, so every report came back as the incomplete fallback.
-    // Object closure is enforced on the response instead, by the `unknown
-    // field` arm of `validate_report_candidate`. Array and numeric bounds are
-    // in the subset and are still sent.
-    let text = || serde_json::json!({ "type": "STRING" });
-    let strings = |min: u32, max: u32| {
-        serde_json::json!({
-            "type": "ARRAY", "minItems": min, "maxItems": max,
-            "items": { "type": "STRING" }
-        })
-    };
-    let feedback = serde_json::json!({
-        "type": "OBJECT",
-        "propertyOrdering": ["strengths", "improvements"],
-        "properties": { "strengths": strings(2, 4), "improvements": strings(2, 4) },
-        "required": ["strengths", "improvements"]
-    });
-    let plan = serde_json::json!({
-        "type": "OBJECT",
-        "propertyOrdering": ["phase", "weakness", "impact", "frequency", "drill", "durationMin", "successCriterion", "selfReview"],
-        "properties": {
-            "phase": { "type": "STRING", "enum": IMPROVEMENT_PHASES },
-            "weakness": text(), "impact": { "type": "STRING", "enum": ["high", "medium", "low"] },
-            "frequency": { "type": "INTEGER", "minimum": 1, "maximum": 99 },
-            "drill": text(), "durationMin": { "type": "INTEGER", "minimum": 1, "maximum": 30 },
-            "successCriterion": text(), "selfReview": strings(1, 4)
-        },
-        "required": ["phase", "weakness", "impact", "frequency", "drill", "durationMin", "successCriterion", "selfReview"]
-    });
-    let assessment_row = serde_json::json!({
-        "type": "OBJECT", "propertyOrdering": ["phase", "score", "weaknessTags"],
-        "properties": {
-            "phase": { "type": "STRING", "enum": IMPROVEMENT_PHASES },
-            "score": { "type": "INTEGER", "minimum": 0, "maximum": 100, "nullable": true },
-            "weaknessTags": strings(0, 4)
-        },
-        "required": ["phase", "score", "weaknessTags"]
-    });
-    serde_json::json!({
-        "type": "OBJECT",
-        "propertyOrdering": ["codingScore", "communicationScore", "decision", "summary", "codingFeedback", "communicationFeedback", "improvementPlan", "frameworkAssessment"],
-        "properties": {
-            "codingScore": { "type": "INTEGER", "minimum": 0, "maximum": 100 },
-            "communicationScore": { "type": "INTEGER", "minimum": 0, "maximum": 100 },
-            "decision": { "type": "STRING", "enum": ["HIRE", "NO_HIRE"] },
-            "summary": text(), "codingFeedback": feedback.clone(), "communicationFeedback": feedback,
-            "improvementPlan": { "type": "ARRAY", "minItems": 0, "maxItems": 8, "items": plan },
-            "frameworkAssessment": {
-                "type": "OBJECT", "propertyOrdering": ["rubricVersion", "phases"],
-                "properties": {
-                    // No `enum` here. Gemini's subset types `Schema.enum` as
-                    // repeated string, so an integer entry fails the whole
-                    // request with "Invalid value ... (TYPE_STRING)". The exact
-                    // value is pinned on the response by `strict_integer`.
-                    "rubricVersion": {
-                        "type": "INTEGER",
-                        "minimum": RUBRIC_VERSION,
-                        "maximum": RUBRIC_VERSION
-                    },
-                    "phases": { "type": "ARRAY", "minItems": 10, "maxItems": 10, "items": assessment_row }
-                },
-                "required": ["rubricVersion", "phases"]
-            }
-        },
-        "required": ["codingScore", "communicationScore", "decision", "summary", "codingFeedback", "communicationFeedback", "improvementPlan", "frameworkAssessment"]
-    })
-}
-
-pub fn validate_report(
-    raw: &serde_json::Value,
-    hints_used: u32,
-) -> Result<serde_json::Value, Vec<String>> {
-    let mut report = validate_report_candidate(raw)?;
-    report
-        .as_object_mut()
-        .expect("validated object")
-        .insert("hintsUsed".to_string(), serde_json::json!(hints_used));
-    Ok(report)
-}
-
-pub fn validate_report_candidate(
-    raw: &serde_json::Value,
-) -> Result<serde_json::Value, Vec<String>> {
-    let mut errors = Vec::new();
-    let Some(object) = raw.as_object() else {
-        return Err(vec!["$: expected object".to_string()]);
-    };
-    exact_keys(
-        object,
-        &[
-            "codingScore",
-            "communicationScore",
-            "decision",
-            "summary",
-            "codingFeedback",
-            "communicationFeedback",
-            "improvementPlan",
-            "frameworkAssessment",
-        ],
-        "$",
-        &mut errors,
-    );
-    strict_integer(
-        object.get("codingScore"),
-        0,
-        100,
-        "$.codingScore",
-        &mut errors,
-    );
-    strict_integer(
-        object.get("communicationScore"),
-        0,
-        100,
-        "$.communicationScore",
-        &mut errors,
-    );
-    strict_enum(
-        object.get("decision"),
-        &["HIRE", "NO_HIRE"],
-        "$.decision",
-        &mut errors,
-    );
-    strict_text(object.get("summary"), 1200, "$.summary", &mut errors);
-    for key in ["codingFeedback", "communicationFeedback"] {
-        validate_feedback(object.get(key), &format!("$.{key}"), &mut errors);
-    }
-    let improvements = object
-        .get("codingFeedback")
-        .and_then(|v| v.get("improvements"))
-        .and_then(serde_json::Value::as_array)
-        .into_iter()
-        .flatten()
-        .chain(
-            object
-                .get("communicationFeedback")
-                .and_then(|v| v.get("improvements"))
-                .and_then(serde_json::Value::as_array)
-                .into_iter()
-                .flatten(),
-        )
-        .filter_map(serde_json::Value::as_str)
-        .map(str::trim)
-        .collect::<Vec<_>>();
-    validate_improvement_plan(object.get("improvementPlan"), &improvements, &mut errors);
-    validate_framework_assessment(
-        object.get("frameworkAssessment"),
-        object.get("improvementPlan"),
-        &mut errors,
-    );
-    for key in [
-        "summary",
-        "codingFeedback",
-        "communicationFeedback",
-        "improvementPlan",
-    ] {
-        if let Some(value) = object.get(key) {
-            validate_observable_judgments(value, &format!("$.{key}"), &mut errors);
-        }
-    }
-    if !errors.is_empty() {
-        return Err(errors);
-    }
-    Ok(raw.clone())
-}
-
-fn validate_observable_judgments(value: &serde_json::Value, path: &str, errors: &mut Vec<String>) {
-    match value {
-        serde_json::Value::String(text) => {
-            let normalized = text
-                .chars()
-                .map(|character| {
-                    if character.is_alphanumeric() {
-                        character.to_ascii_lowercase()
-                    } else {
-                        ' '
-                    }
-                })
-                .collect::<String>()
-                .split_whitespace()
-                .collect::<Vec<_>>()
-                .join(" ");
-            let padded = format!(" {normalized} ");
-            const UNSUPPORTED: &[&str] = &[
-                " accent ",
-                " accents ",
-                " dialect ",
-                " dialects ",
-                " typing speed ",
-                " typing pace ",
-                " typed slowly ",
-                " typed quickly ",
-                " type slowly ",
-                " type quickly ",
-                " speech rate ",
-                " filler word ",
-                " filler words ",
-                " disfluency ",
-                " disfluencies ",
-                " eye contact ",
-                " posture ",
-                " body language ",
-                " facial expression ",
-                " facial expressions ",
-                " voice tone ",
-                " vocal tone ",
-                " tone of voice ",
-                " attractiveness ",
-                " physical appearance ",
-                " nervous demeanor ",
-                " confident demeanor ",
-                " personality ",
-                " introvert ",
-                " extrovert ",
-                " charisma ",
-                " appeared nervous ",
-                " appears nervous ",
-                " seemed nervous ",
-                " looked nervous ",
-                " sounded nervous ",
-                " come across as nervous ",
-                " comes across as nervous ",
-                " appeared confident ",
-                " appears confident ",
-                " seemed confident ",
-                " looked confident ",
-                " sounded confident ",
-                " come across as confident ",
-                " comes across as confident ",
-            ];
-            if UNSUPPORTED.iter().any(|phrase| padded.contains(phrase)) {
-                errors.push(format!(
-                    "{path}: unsupported delivery or personality judgment"
-                ));
-            }
-        }
-        serde_json::Value::Array(items) => {
-            for (index, item) in items.iter().enumerate() {
-                validate_observable_judgments(item, &format!("{path}[{index}]"), errors);
-            }
-        }
-        serde_json::Value::Object(object) => {
-            for (key, item) in object {
-                validate_observable_judgments(item, &format!("{path}.{key}"), errors);
-            }
-        }
-        _ => {}
-    }
-}
-
-fn exact_keys(
-    object: &serde_json::Map<String, serde_json::Value>,
-    expected: &[&str],
-    path: &str,
-    errors: &mut Vec<String>,
-) {
-    for key in expected {
-        if !object.contains_key(*key) {
-            errors.push(format!("{path}.{key}: required"));
-        }
-    }
-    for key in object.keys() {
-        if !expected.contains(&key.as_str()) {
-            errors.push(format!("{path}.{key}: unknown field"));
-        }
-    }
-}
-
-fn strict_integer(
-    value: Option<&serde_json::Value>,
-    min: i64,
-    max: i64,
-    path: &str,
-    errors: &mut Vec<String>,
-) -> Option<i64> {
-    match value
-        .and_then(serde_json::Value::as_i64)
-        .filter(|v| (min..=max).contains(v))
-    {
-        Some(value) => Some(value),
-        None => {
-            errors.push(format!("{path}: expected integer {min}..={max}"));
-            None
-        }
-    }
-}
-
-fn strict_enum<'a>(
-    value: Option<&'a serde_json::Value>,
-    allowed: &[&str],
-    path: &str,
-    errors: &mut Vec<String>,
-) -> Option<&'a str> {
-    match value
-        .and_then(serde_json::Value::as_str)
-        .filter(|v| allowed.contains(v))
-    {
-        Some(value) => Some(value),
-        None => {
-            errors.push(format!("{path}: invalid enum"));
-            None
-        }
-    }
-}
-
-fn strict_text(
-    value: Option<&serde_json::Value>,
-    max: usize,
-    path: &str,
-    errors: &mut Vec<String>,
-) -> Option<String> {
-    match value
-        .and_then(serde_json::Value::as_str)
-        .filter(|value| !value.trim().is_empty() && value.chars().count() <= max)
-    {
-        Some(value) => Some(value.trim().to_string()),
-        None => {
-            errors.push(format!(
-                "{path}: expected non-empty string of at most {max} characters"
-            ));
-            None
-        }
-    }
-}
-
-fn validate_string_array(
-    value: Option<&serde_json::Value>,
-    min: usize,
-    max: usize,
-    text_max: usize,
-    path: &str,
-    errors: &mut Vec<String>,
-) {
-    let Some(items) = value.and_then(serde_json::Value::as_array) else {
-        errors.push(format!("{path}: expected array"));
-        return;
-    };
-    if !(min..=max).contains(&items.len()) {
-        errors.push(format!("{path}: expected {min}..={max} items"));
-    }
-    let mut seen = std::collections::HashSet::new();
-    for (index, item) in items.iter().take(max.saturating_add(1)).enumerate() {
-        if let Some(text) = strict_text(Some(item), text_max, &format!("{path}[{index}]"), errors)
-            && !seen.insert(text)
-        {
-            errors.push(format!("{path}[{index}]: duplicate"));
-        }
-    }
-}
-
-fn validate_feedback(value: Option<&serde_json::Value>, path: &str, errors: &mut Vec<String>) {
-    let Some(object) = value.and_then(serde_json::Value::as_object) else {
-        errors.push(format!("{path}: expected object"));
-        return;
-    };
-    exact_keys(object, &["strengths", "improvements"], path, errors);
-    validate_string_array(
-        object.get("strengths"),
-        2,
-        4,
-        400,
-        &format!("{path}.strengths"),
-        errors,
-    );
-    validate_string_array(
-        object.get("improvements"),
-        2,
-        4,
-        400,
-        &format!("{path}.improvements"),
-        errors,
-    );
-}
-
-fn validate_improvement_plan(
-    value: Option<&serde_json::Value>,
-    weaknesses: &[&str],
-    errors: &mut Vec<String>,
-) {
-    let Some(items) = value.and_then(serde_json::Value::as_array) else {
-        errors.push("$.improvementPlan: expected array".to_string());
-        return;
-    };
-    if items.len() > 8 {
-        errors.push("$.improvementPlan: expected at most 8 items".to_string());
-    }
-    let mut seen = std::collections::HashSet::new();
-    let mut previous_order = None;
-    for (index, item) in items.iter().take(9).enumerate() {
-        let path = format!("$.improvementPlan[{index}]");
-        let Some(object) = item.as_object() else {
-            errors.push(format!("{path}: expected object"));
-            continue;
-        };
-        exact_keys(
-            object,
-            &[
-                "phase",
-                "weakness",
-                "impact",
-                "frequency",
-                "drill",
-                "durationMin",
-                "successCriterion",
-                "selfReview",
-            ],
-            &path,
-            errors,
-        );
-        strict_enum(
-            object.get("phase"),
-            &IMPROVEMENT_PHASES,
-            &format!("{path}.phase"),
-            errors,
-        );
-        let weakness = strict_text(
-            object.get("weakness"),
-            400,
-            &format!("{path}.weakness"),
-            errors,
-        );
-        if let Some(weakness) = weakness {
-            if !weaknesses.contains(&weakness.as_str()) {
-                errors.push(format!(
-                    "{path}.weakness: must exactly reference a feedback improvement"
-                ));
-            }
-            if !seen.insert(weakness) {
-                errors.push(format!("{path}.weakness: duplicate"));
-            }
-        }
-        let impact = strict_enum(
-            object.get("impact"),
-            &["high", "medium", "low"],
-            &format!("{path}.impact"),
-            errors,
-        );
-        let frequency = strict_integer(
-            object.get("frequency"),
-            1,
-            99,
-            &format!("{path}.frequency"),
-            errors,
-        );
-        if let (Some(impact), Some(frequency)) = (impact, frequency) {
-            let rank = match impact {
-                "high" => 3,
-                "medium" => 2,
-                _ => 1,
-            };
-            let order = (rank, frequency);
-            if previous_order.is_some_and(|previous| previous < order) {
-                errors.push(format!(
-                    "{path}: items must be sorted by impact then frequency descending"
-                ));
-            }
-            previous_order = Some(order);
-        }
-        strict_text(object.get("drill"), 400, &format!("{path}.drill"), errors);
-        strict_integer(
-            object.get("durationMin"),
-            1,
-            30,
-            &format!("{path}.durationMin"),
-            errors,
-        );
-        strict_text(
-            object.get("successCriterion"),
-            400,
-            &format!("{path}.successCriterion"),
-            errors,
-        );
-        validate_string_array(
-            object.get("selfReview"),
-            1,
-            4,
-            240,
-            &format!("{path}.selfReview"),
-            errors,
-        );
-    }
-    let expected = weaknesses
-        .iter()
-        .copied()
-        .collect::<std::collections::HashSet<_>>();
-    let actual = seen
-        .iter()
-        .map(String::as_str)
-        .collect::<std::collections::HashSet<_>>();
-    if actual != expected {
-        errors.push(
-            "$.improvementPlan: must contain exactly one item for every feedback improvement"
-                .to_string(),
-        );
-    }
-}
-
-fn validate_framework_assessment(
-    value: Option<&serde_json::Value>,
-    plan: Option<&serde_json::Value>,
-    errors: &mut Vec<String>,
-) {
-    let Some(object) = value.and_then(serde_json::Value::as_object) else {
-        errors.push("$.frameworkAssessment: expected object".to_string());
-        return;
-    };
-    exact_keys(
-        object,
-        &["rubricVersion", "phases"],
-        "$.frameworkAssessment",
-        errors,
-    );
-    strict_integer(
-        object.get("rubricVersion"),
-        i64::from(RUBRIC_VERSION),
-        i64::from(RUBRIC_VERSION),
-        "$.frameworkAssessment.rubricVersion",
-        errors,
-    );
-    let Some(rows) = object.get("phases").and_then(serde_json::Value::as_array) else {
-        errors.push("$.frameworkAssessment.phases: expected array".to_string());
-        return;
-    };
-    if rows.len() != IMPROVEMENT_PHASES.len() {
-        errors.push("$.frameworkAssessment.phases: expected exactly 10 ordered phases".to_string());
-    }
-    let plan = plan
-        .and_then(serde_json::Value::as_array)
-        .map(Vec::as_slice)
-        .unwrap_or(&[]);
-    for (index, expected_phase) in IMPROVEMENT_PHASES.iter().enumerate() {
-        let path = format!("$.frameworkAssessment.phases[{index}]");
-        let Some(row) = rows.get(index).and_then(serde_json::Value::as_object) else {
-            errors.push(format!("{path}: expected object"));
-            continue;
-        };
-        exact_keys(row, &["phase", "score", "weaknessTags"], &path, errors);
-        if row.get("phase").and_then(serde_json::Value::as_str) != Some(*expected_phase) {
-            errors.push(format!("{path}.phase: expected {expected_phase}"));
-        }
-        if row.get("score") != Some(&serde_json::Value::Null) {
-            strict_integer(row.get("score"), 0, 100, &format!("{path}.score"), errors);
-        }
-        validate_string_array(
-            row.get("weaknessTags"),
-            0,
-            4,
-            400,
-            &format!("{path}.weaknessTags"),
-            errors,
-        );
-        if let Some(tags) = row
-            .get("weaknessTags")
-            .and_then(serde_json::Value::as_array)
-        {
-            // Trimmed on both sides, like every other comparison in this
-            // module: `validate_improvement_plan` matches a plan weakness to a
-            // feedback improvement after trimming, so a tag that is an exact
-            // copy of an accepted weakness apart from surrounding whitespace
-            // has to be accepted here too. Comparing raw rejected the whole
-            // report over a trailing space and spent the one repair on it.
-            let allowed = plan
-                .iter()
-                .filter(|item| {
-                    item.get("phase").and_then(serde_json::Value::as_str) == Some(*expected_phase)
-                })
-                .filter_map(|item| item.get("weakness").and_then(serde_json::Value::as_str))
-                .map(str::trim)
-                .collect::<std::collections::HashSet<_>>();
-            for tag in tags.iter().filter_map(serde_json::Value::as_str) {
-                if !allowed.contains(tag.trim()) {
-                    errors.push(format!(
-                        "{path}.weaknessTags: tag has no same-phase improvement"
-                    ));
-                }
-            }
-        }
-    }
-}
-
-const IMPROVEMENT_PHASES: [&str; 10] = [
-    "Repeat",
-    "Example",
-    "Algorithm",
-    "Coding",
-    "Test",
-    "Optimizations",
-    "Situation",
-    "Task",
-    "Action",
-    "Result",
-];
-
 pub fn spoken_minutes_from_remaining_seconds(remaining_seconds: i64) -> i64 {
     ((remaining_seconds as f64) / 60.0)
         .round_ties_even()
@@ -1610,427 +1006,6 @@ pub fn test_reaction_decision(
     }
 }
 
-pub fn apply_data_event(
-    state: &mut RuntimeState,
-    topic: &str,
-    payload: &serde_json::Value,
-    since_last_test_reaction_seconds: f64,
-) -> DataEventResult {
-    if state.behavioral_round_started && matches!(topic, TOPIC_CODE_UPDATE | TOPIC_TEST_RESULTS) {
-        return DataEventResult::default();
-    }
-    if state.paused && matches!(topic, TOPIC_CODE_UPDATE | TOPIC_TEST_RESULTS) {
-        return DataEventResult::default();
-    }
-    match topic {
-        TOPIC_CODE_UPDATE => apply_code_update(state, payload),
-        TOPIC_TEST_RESULTS => apply_test_results(state, payload, since_last_test_reaction_seconds),
-        TOPIC_CONTROL => apply_control(state, payload),
-        TOPIC_INTEGRITY => apply_integrity(state, payload),
-        _ => DataEventResult::default(),
-    }
-}
-
-/// The report, or a stub saying why there is none.
-///
-/// Every fallback is logged, because the note it carries reaches the candidate
-/// and nobody else. A schema the provider refused answered 400 on every
-/// interview for the life of a branch while the suite stayed green, and the
-/// only place that was visible was a sentence in the candidate's own report.
-/// An operator reading the log sees it now, on the line that produced it, with
-/// the validation errors rather than only the fact that there were some: which
-/// field the model got wrong is the whole diagnostic, and the stub the
-/// candidate receives cannot carry it.
-pub fn final_report(
-    raw_report: Option<&serde_json::Value>,
-    hints_used: u32,
-    error_note: Option<&str>,
-) -> serde_json::Value {
-    let (reason, errors) = match (raw_report, error_note) {
-        (Some(raw_report), None) => match validate_report(raw_report, hints_used) {
-            Ok(report) => return report,
-            Err(errors) => ("report schema validation failed", errors),
-        },
-        _ => (error_note.unwrap_or("report generation failed"), Vec::new()),
-    };
-    eprintln!(
-        "codetrial report_incomplete reason={reason}{}",
-        // Bounded, because a wholly wrong shape produces one per field and this
-        // is a log line, not the report.
-        errors
-            .iter()
-            .take(5)
-            .map(|error| format!(" | {error}"))
-            .collect::<String>()
-    );
-    fallback_report(hints_used, reason)
-}
-
-/// The id and its spoken form, or nothing at all when the browser sent
-/// something the tabs do not offer.
-fn offered_language(language: &str) -> Option<(&str, &'static str)> {
-    spoken_language(language).map(|spoken| (language, spoken))
-}
-
-fn apply_code_update(state: &mut RuntimeState, payload: &serde_json::Value) -> DataEventResult {
-    // A packet with no string `code` is a malformed packet, not an empty
-    // editor. Defaulting to "" meant one of those wiped the authoritative
-    // buffer, and the buffer is what the report is written from.
-    let new_code = payload.get("code").and_then(serde_json::Value::as_str);
-
-    // Editor packets arrive several times a second and are usually identical to
-    // the last one, so only touch the buffer when the text actually moved.
-    let first_code_packet = state.code.is_empty();
-    let update_last_code_change = new_code.is_some_and(|code| code != state.code);
-    if let Some(code) = new_code.filter(|_| update_last_code_change) {
-        state.code.clear();
-        state.code.push_str(code);
-    }
-
-    // A language switch is silent from the interviewer's side: the click swaps
-    // the editor buffer and publishes the same code topic as any keystroke. The
-    // greeting asks the candidate to pick one, so leaving the pick
-    // unacknowledged makes the interviewer look like they missed it.
-    //
-    // Only a real change speaks. The browser publishes its starting language on
-    // connect, which matches the default and must not produce a confirmation
-    // for a choice nobody made. Only a language the tabs actually offer may
-    // change the state or reach a prompt. The id comes from the candidate's
-    // browser and is interpolated into Gemini's instructions, so an unvalidated
-    // one is a way to write into them.
-    let switching = payload
-        .get("language")
-        .and_then(serde_json::Value::as_str)
-        .filter(|language| *language != state.language)
-        .and_then(offered_language);
-
-    // A non-empty buffer is not evidence that the candidate has done anything.
-    // The browser publishes the starter template on connect and publishes the
-    // next language's template on every switch, in the same packet as the
-    // switch, so the buffer is a template from the first second of the
-    // interview onward. What separates work from a template is a change that
-    // arrived without a language switch, which is a keystroke.
-    if update_last_code_change && !first_code_packet && switching.is_none() {
-        state.code_edited = true;
-    }
-
-    let mut language_changed = None;
-    if let Some(spoken) = switching {
-        state.language = spoken.0.to_string();
-
-        // Reading the buffer here instead asked Gemini not to make a candidate
-        // "restate work they already completed" for a template they had not
-        // touched, which skipped the restatement the whole REACTO opening is
-        // built on for anyone who picked a language before speaking.
-        let context = if state.code_edited {
-            LanguageChoiceContext::SwitchWithCode
-        } else {
-            LanguageChoiceContext::Start
-        };
-        language_changed = Some(language_choice(spoken.1, context));
-    }
-
-    DataEventResult {
-        update_last_code_change,
-
-        // Counts as an interjection: it is the interviewer speaking unprompted,
-        // and without this a candidate trying two tabs in a row is talked over
-        // twice while the cooldown thinks nothing has been said.
-        update_last_interjection: language_changed.is_some(),
-        generate_reply: language_changed,
-        ..DataEventResult::default()
-    }
-}
-
-/// # What a passing run means
-///
-/// Nothing, on its own. `passed` and `total` are counted in the candidate's
-/// browser against a judge the candidate's browser holds, and
-/// `agent/integrity.rs` already says why that makes them an account rather than
-/// evidence: the reporter is the adversary. Believing the count here is a
-/// deliberate product choice, not an oversight. The interviewer reacts to what
-/// the candidate says happened, the same way a human interviewer reacts to
-/// "that one passes" without rerunning it.
-///
-/// So the numbers are safe to narrate and unsafe to score. What carries weight
-/// is the code itself, which the agent receives over the code topic and can
-/// read, and the room state it observes for itself. Anything that grades rather
-/// than converses has to judge server-side, against code the server holds.
-fn apply_test_results(
-    state: &mut RuntimeState,
-    payload: &serde_json::Value,
-    since_last_test_reaction_seconds: f64,
-) -> DataEventResult {
-    // Bounded before it is stored, not before it is rendered. Both readers of
-    // `last_test_run` put it in front of a model, so the sanitized value has to
-    // be the only one that exists past this line.
-    let payload = &sanitize_test_run(payload);
-
-    // The sanitizer reports "no run happened" as null, and the caller has to
-    // honor that or the refusal to invent a run is undone one line later:
-    // counting it and reacting to it is what telling the report a run occurred
-    // looks like from here.
-    if payload.is_null() {
-        return DataEventResult::default();
-    }
-    state.last_test_run = Some(payload.clone());
-    state.test_runs += 1;
-
-    let decision = test_reaction_decision(state.ended, since_last_test_reaction_seconds);
-    if !decision.react {
-        return DataEventResult::default();
-    }
-
-    let all_passed = payload
-        .get("setupError")
-        .is_none_or(|value| !python_truthy(value))
-        && payload
-            .get("total")
-            .and_then(serde_json::Value::as_i64)
-            .is_some_and(|total| total > 0)
-        && payload.get("passed").and_then(serde_json::Value::as_i64)
-            == payload.get("total").and_then(serde_json::Value::as_i64);
-    let summary = format_test_run(Some(payload), state.test_runs);
-
-    DataEventResult {
-        update_last_code_change: false,
-        update_last_test_reaction: true,
-        update_last_interjection: true,
-        generate_reply: Some(test_results_reaction(&summary, all_passed)),
-        finish_interview: None,
-        ..DataEventResult::default()
-    }
-}
-
-fn apply_control(state: &mut RuntimeState, payload: &serde_json::Value) -> DataEventResult {
-    match payload.get("type").and_then(serde_json::Value::as_str) {
-        // Pause used to be a practice-only affordance, and practice is gone.
-        // Left working for everyone rather than deleted with the mode: a
-        // candidate whose machine or network interrupts them mid-interview can
-        // at least stop the interviewer talking into an empty room. It stops
-        // the conversation, not the deadline, which runs on wall clock either
-        // way. It is recorded, so a paused stretch is visible in the report.
-        Some("pause_interview") if !state.ended => {
-            let paused = payload
-                .get("paused")
-                .and_then(serde_json::Value::as_bool)
-                .unwrap_or(false);
-            if paused == state.paused {
-                return DataEventResult::default();
-            }
-            state.paused = paused;
-            DataEventResult {
-                pause_changed: Some(paused),
-                generate_reply: (!paused).then(|| {
-                    "The interview has resumed. Continue with your REACTO step.".to_string()
-                }),
-                ..DataEventResult::default()
-            }
-        }
-        Some("round_transition")
-            if !state.ended
-                && !state.paused
-                && state.interview_loop == InterviewLoop::CodingBehavioral
-                && !state.round_transition_seen
-                && payload.get("round").and_then(serde_json::Value::as_str)
-                    == Some("behavioral")
-                && state.started_at.elapsed() + ROUND_TRANSITION_SKEW
-                    >= std::time::Duration::from_secs(u64::from(state.coding_minutes) * 60) =>
-        {
-            state.round_transition_seen = true;
-            let completed = |phase| {
-                state
-                    .framework_evidence
-                    .iter()
-                    .any(|item| item.phase == phase && item.kind != EvidenceKind::Skipped)
-            };
-            if completed(FrameworkPhase::Test) && completed(FrameworkPhase::Optimizations) {
-                state.behavioral_round_started = true;
-                DataEventResult {
-                    round_changed: Some("started"),
-                    generate_reply: Some("[SYSTEM EVENT] The trusted coding completion gate passed: Test and Optimizations both have candidate evidence. The coding round is closed. Begin the reserved behavioral round now with exactly one concise question under the private STAR, profile, and document-grounding policies. Use prior candidate answers only to deepen the follow-up; do not repeat them and do not return to coding.".to_string()),
-                    ..DataEventResult::default()
-                }
-            } else {
-                DataEventResult {
-                    round_changed: Some("skipped"),
-                    generate_reply: Some("[SYSTEM EVENT] The behavioral reserve began, but the trusted coding completion gate did not pass because Test or Optimizations evidence is absent. Do not start STAR. Keep the candidate focused on a testable solution, highest-value tests, and justified complexity until the session ends; missing STAR phases will be marked skipped.".to_string()),
-                    ..DataEventResult::default()
-                }
-            }
-        }
-        Some("time_warning") if !state.ended && !state.paused => {
-            let remaining_seconds = payload
-                .get("remainingSeconds")
-                .and_then(json_int)
-                .unwrap_or(300);
-            let minutes = spoken_minutes_from_remaining_seconds(remaining_seconds) as u32;
-            DataEventResult {
-                generate_reply: Some(if state.behavioral_round_started {
-                    format!(
-                        "[SYSTEM EVENT] Exactly {minutes} minutes remain in the active behavioral round. Do not return to coding or ask a new question. Let the candidate finish the current answer, ask at most the one permitted neutral missing-STAR follow-up, then close naturally."
-                    )
-                } else {
-                    time_warning(minutes)
-                }),
-                ..DataEventResult::default()
-            }
-        }
-        Some("end_interview") if !state.ended => {
-            if !state.behavioral_round_started
-                && !payload.get("code").is_some_and(serde_json::Value::is_null)
-                && let Some(code) = payload.get("code").and_then(serde_json::Value::as_str)
-            {
-                state.code = code.to_string();
-            }
-
-            // Read independently of the code. Nesting it meant a payload whose
-            // code was absent or null also discarded the language, and the
-            // report prompt was then written against whatever language the
-            // session last happened to record. Validated for the same reason as
-            // the code path: this string reaches the report prompt.
-            if !state.behavioral_round_started
-                && let Some((id, _)) = payload
-                    .get("language")
-                    .and_then(serde_json::Value::as_str)
-                    .and_then(offered_language)
-            {
-                state.language = id.to_string();
-            }
-            state.ended = true;
-            DataEventResult {
-                finish_interview: Some(
-                    payload
-                        .get("reason")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("candidate_ended")
-                        .to_string(),
-                ),
-                ..DataEventResult::default()
-            }
-        }
-        _ => DataEventResult::default(),
-    }
-}
-
-fn apply_integrity(state: &mut RuntimeState, payload: &serde_json::Value) -> DataEventResult {
-    let Some(event) = sanitize_integrity_event(payload) else {
-        return DataEventResult::default();
-    };
-    let seq = event.get("seq").and_then(serde_json::Value::as_u64);
-    let prev_hash = event.get("prevHash").and_then(serde_json::Value::as_str);
-    let (expected_seq, expected_prev_hash) = state
-        .integrity_chain
-        .as_ref()
-        .map_or((1, ""), |(seq, hash)| {
-            (seq.saturating_add(1), hash.as_str())
-        });
-
-    let computed_hash = integrity_hash(&event);
-
-    // Named, because a rejection here is permanent. Sequence continuity is
-    // required, so `expected_seq` never advances past a refused event and every
-    // later one is refused for the same reason. Four different conditions used
-    // to return the same silent default, and a hash mismatch caused by this
-    // process normalizing `detail` before rehashing it emptied the evidence
-    // section of every camera interview without a line anywhere saying so.
-    //
-    // Only tampering belongs here. A retention decision is handled below, after
-    // the cursor has moved, because punishing the producer for this process
-    // running out of room is how a full buffer used to end the chain.
-    let refusal = if seq != Some(expected_seq) {
-        Some(format!("sequence expected {expected_seq}, got {seq:?}"))
-    } else if prev_hash != Some(expected_prev_hash) {
-        Some("previous hash does not match the last accepted event".to_string())
-    } else if event.get("hash").and_then(serde_json::Value::as_str) != Some(computed_hash.as_str())
-    {
-        Some("hash mismatch; the producer and this verifier disagree".to_string())
-    } else {
-        None
-    };
-    if let Some(reason) = refusal {
-        eprintln!(
-            "WARNING: integrity event refused, chain stops here: {reason} (seq={seq:?}, type={})",
-            event
-                .get("type")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or("?")
-        );
-        return DataEventResult::default();
-    }
-
-    // The chain is satisfied, so the cursor moves whatever happens to the event
-    // itself. Everything past this line is about what is worth keeping, and
-    // none of it can refuse the next event.
-    //
-    // Built from the values the checks above already proved equal, rather than
-    // re-read from the event. A conditional here would have a branch that
-    // silently leaves the cursor behind, which is the failure this separation
-    // exists to remove.
-    state.integrity_chain = Some((expected_seq, computed_hash));
-
-    // Liveness, not evidence, and now stored as such. Keeping only the ends of
-    // the run is what lets the evidence buffer be a plain queue again.
-    if is_heartbeat(&event) {
-        if state.integrity_first_heartbeat.is_none() {
-            state.integrity_first_heartbeat = Some(event);
-        } else {
-            state.integrity_last_heartbeat = Some(event);
-        }
-        return DataEventResult::default();
-    }
-
-    if !valid_review_sources(state, &event) {
-        eprintln!(
-            "WARNING: integrity event dropped from the evidence, chain continues: review event \
-             references an event no longer retained (seq={seq:?})"
-        );
-        return DataEventResult::default();
-    }
-
-    state.integrity_events.push(event);
-
-    // Plain oldest-out, because nothing periodic lands here any more. This used
-    // to hold heartbeats too, and at one every five seconds they filled all 25
-    // slots in about two minutes: a camera that stopped at minute three was
-    // gone from a thirty minute report by minute five.
-    if state.integrity_events.len() > MAX_INTEGRITY_EVENTS {
-        let excess = state.integrity_events.len() - MAX_INTEGRITY_EVENTS;
-        state.integrity_events.drain(0..excess);
-    }
-    DataEventResult::default()
-}
-
-/// Liveness rather than evidence, which is what sends it to its own field
-/// instead of competing for a slot in the buffer.
-fn is_heartbeat(event: &serde_json::Value) -> bool {
-    event.get("type").and_then(serde_json::Value::as_str) == Some("INTEGRITY_HEARTBEAT")
-}
-
-pub(crate) fn valid_review_sources(state: &RuntimeState, event: &serde_json::Value) -> bool {
-    if event.get("type").and_then(serde_json::Value::as_str) != Some("REVIEW_EVENT") {
-        return true;
-    }
-    let Some(ids) = event
-        .get("sourceEventIds")
-        .and_then(serde_json::Value::as_array)
-    else {
-        return false;
-    };
-    if ids.is_empty() {
-        return false;
-    }
-    ids.iter().all(|id| {
-        let Some(seq) = id.as_str().and_then(|value| value.parse::<u64>().ok()) else {
-            return false;
-        };
-        state
-            .integrity_events
-            .iter()
-            .any(|stored| stored.get("seq").and_then(serde_json::Value::as_u64) == Some(seq))
-    })
-}
-
 pub(crate) fn duration_from_metadata(value: Option<&serde_json::Value>) -> u32 {
     value
         .filter(|value| python_truthy(value))
@@ -2038,117 +1013,4 @@ pub(crate) fn duration_from_metadata(value: Option<&serde_json::Value>) -> u32 {
         .map_or(DEFAULT_DURATION_MIN, |duration| {
             duration.clamp(MIN_DURATION_MIN.into(), MAX_DURATION_MIN.into()) as u32
         })
-}
-
-/// Python-compatible coercion of a JSON scalar to a number. The browser has
-/// historically sent durations and scores as numbers, numeric strings, and
-/// booleans, so both the token endpoint and the agent decode them this way.
-pub fn json_number(value: &serde_json::Value) -> Option<f64> {
-    match value {
-        serde_json::Value::Number(number) => number.as_f64(),
-        serde_json::Value::String(text) => text.trim().parse().ok(),
-        serde_json::Value::Bool(value) => Some(f64::from(*value)),
-        _ => None,
-    }
-}
-
-pub(crate) fn json_int(value: &serde_json::Value) -> Option<i64> {
-    json_number(value).map(|number| number as i64)
-}
-
-pub(crate) fn truthy_string(value: Option<&serde_json::Value>) -> Option<String> {
-    value
-        .filter(|value| python_truthy(value))
-        .map(python_string)
-}
-
-pub(crate) fn value_string(value: Option<&serde_json::Value>) -> Option<String> {
-    value.map(python_string)
-}
-
-pub(crate) fn python_truthy(value: &serde_json::Value) -> bool {
-    match value {
-        serde_json::Value::Null => false,
-        serde_json::Value::Bool(value) => *value,
-        serde_json::Value::Number(number) => number.as_f64() != Some(0.0),
-        serde_json::Value::String(text) => !text.is_empty(),
-        serde_json::Value::Array(items) => !items.is_empty(),
-        serde_json::Value::Object(fields) => !fields.is_empty(),
-    }
-}
-
-pub(crate) fn python_string(value: &serde_json::Value) -> String {
-    match value {
-        serde_json::Value::Null => "None".to_string(),
-        serde_json::Value::Bool(true) => "True".to_string(),
-        serde_json::Value::Bool(false) => "False".to_string(),
-        serde_json::Value::Number(number) => number.to_string(),
-        serde_json::Value::String(text) => text.clone(),
-        serde_json::Value::Array(items) => format!(
-            "[{}]",
-            items.iter().map(python_repr).collect::<Vec<_>>().join(", ")
-        ),
-        serde_json::Value::Object(fields) => format!(
-            "{{{}}}",
-            fields
-                .iter()
-                .map(|(key, value)| format!("'{key}': {}", python_repr(value)))
-                .collect::<Vec<_>>()
-                .join(", ")
-        ),
-    }
-}
-
-pub(crate) fn python_repr(value: &serde_json::Value) -> String {
-    match value {
-        serde_json::Value::String(text) => format!("'{text}'"),
-        value => python_string(value),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{python_repr, python_string, python_truthy};
-    use serde_json::json;
-
-    /// These render the values a candidate's Python assertion failed on, so an
-    /// empty list has to read as `[]` and false, not as `None` or true. Kept
-    /// here rather than in `tests/`, which would mean making three internal
-    /// helpers public to reach them.
-    #[test]
-    fn python_coercion_matches_the_interpreter() {
-        for falsy in [
-            json!(null),
-            json!(false),
-            json!(0),
-            json!(""),
-            json!([]),
-            json!({}),
-        ] {
-            assert!(!python_truthy(&falsy), "{falsy} should be falsy");
-        }
-        for truthy in [
-            json!(true),
-            json!(42.5),
-            json!("hello"),
-            json!([1]),
-            json!({"k": 1}),
-        ] {
-            assert!(python_truthy(&truthy), "{truthy} should be truthy");
-        }
-
-        assert_eq!(python_string(&json!(null)), "None");
-        assert_eq!(python_string(&json!(true)), "True");
-        assert_eq!(python_string(&json!(false)), "False");
-        assert_eq!(python_string(&json!(42.5)), "42.5");
-        assert_eq!(python_string(&json!("hello")), "hello");
-        assert_eq!(
-            python_string(&json!([1, "a", null, true])),
-            "[1, 'a', None, True]"
-        );
-
-        // The difference that matters: `repr` quotes a string, `str` does not.
-        assert_eq!(python_repr(&json!("hello")), "'hello'");
-        assert_eq!(python_repr(&json!(42.5)), "42.5");
-    }
 }
